throughput_python/requirements.txt
pymongo
profilehooks
redis
sqlalchemy-foundationdb
